Listando todas as tasks de um usuário

+++ backend/src/main/java/com/br/api/backend/model/UserModel.java
@@ -17,7 +17,7 @@ public class UserModel {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long Id;
+    private int Id;
 
     private String name;
